hotfix: reorganized code for clarity; changed handling of generated Company Roles Ref
Company Roles Ref is no longer part of 'all' and has to be made seperately. This is in anticpation for reworking how referenes are generated, which is noted as a TODO in the liq-refs-gen.pl file. Also added a good deal of inline notes in that file to clarify things for future.